[ci skip] Sooooo much change. Making layers more OOP
