+++ smrt/__init__.py
@@ -4,10 +4,7 @@
 #
 # The SMRT module
 
-import os
-import sys
-
-__version__ = '0.1'
+__version__ = '0.2'
 
 try:
     # this var is injected in the setup build to enable
@@ -18,6 +15,8 @@ except NameError:
     __SMRT_SETUP__ = False
 
 if __SMRT_SETUP__:
+    import sys
+    import os
     sys.stderr.write('Partial import of SMRT during the build process.' + os.linesep)
 else:
     __all__ = [
@@ -25,6 +24,14 @@ else:
         'balance'
     ]
 
-    # top-level imports
-    from .balance import smrt_balance, smote_balance
-    from .autoencode.autoencoder import *
+    # top-level imports - if any
+    # todo
+
+
+def setup_module(module):
+    import numpy as np
+    import random
+
+    _random_seed = int(np.random.uniform() * (2 ** 31 - 1))
+    np.random.seed(_random_seed)
+    random.seed(_random_seed)
