+++ smrt/balance/tests/test_smrt.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+#
+# Author: Taylor Smith <taylor.smith@alkaline-ml.com>
+#
+# Test the SMRT balancer
+
+from __future__ import division, absolute_import, division
+
+
+def test_smrt():
+    pass
